fix: update id type

+++ node/migrations/000001_initialize_tables.up.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS feeds (
     id SERIAL PRIMARY KEY,
     name TEXT NOT NULL,
     definition JSONB NOT NULL,
-    adapter_id INT8 NOT NULL,
+    adapter_id INT16 NOT NULL,
     CONSTRAINT feeds_adapter_id_fkey
         FOREIGN KEY(adapter_id)
         REFERENCES adapters(id)
